refactor: move ActionBarButton inside related component
